backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mariadb==1.1.8
PyJWT==2.8.0
